requirements.txt
python==3.6
mxnet-cu92==1.5.0
gluoncv==0.4.0
progress==1.5
